moved ipc test to test proejct
