+++ tests/Tests/Tests.csproj
@@ -6,7 +6,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\RepoM.Api\RepoM.Api.csproj" />
-    <ProjectReference Include="..\..\src\RepoM.Ipc\RepoM.Ipc.csproj" />
   </ItemGroup>
 
   <ItemGroup>
